BE/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0
psycopg2-binary>=2.9.5
django-cors-headers>=3.14.0
google-generativeai